requirements.txt
fpdf2==2.4.6
mysql-connector-python==8.0.27
Pillow==8.4.0